website 1.2 - new sections and improvements
transfered every 'css-thing' from tthe HTML5 to the CSS3 document (to increase performance?).
Also added some new sections with texts.
I also thought of a new version numbering: first number stands for the entire design of the website. If that changes, the number changes as well. The second will change if new thing will be added. There will eventually be a third number at the end, which will be showing smaller changes, like fixing, text editing or smth. Currently we are on 1.3:
1.0: [creation]
1.1:   SEO and PWA (beta)
1.2:  new sections and improvements